Move latest_only_operator.py to latest_only.py (#11178)
