+++ airflow/example_dags/example_latest_only.py
@@ -22,7 +22,7 @@ import datetime as dt
 
 from airflow import DAG
 from airflow.operators.dummy_operator import DummyOperator
-from airflow.operators.latest_only_operator import LatestOnlyOperator
+from airflow.operators.latest_only import LatestOnlyOperator
 from airflow.utils.dates import days_ago
 
 dag = DAG(
